erp/requirements.txt
Django==4.2
djangorestframework==3.14.0
pymysql==1.1.0
django-cors-headers==4.0.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
